# docs/swagger.yaml
basePath: /api/v1/
definitions:
...
  v1.CreateUserRequest:
    properties:
      email:
        description: User email
        type: string
      firstName:
        description: User first name
        type: string
      lastName:
        description: User last name
        type: string
      password:
        description: User password (minimum 8 characters)
        type: string
    type: object
  v1.ErrorResponse:
    properties:
      message:
        type: string
    type: object
...
  v1.UserResponse:
    properties:
      email:
        description: User email
        type: string
      firstName:
        description: User first name
        type: string
      id:
        description: User identifier
        type: integer
      lastName:
        description: User last name
        type: string
    type: object
info:
  contact:
    email: kerhael.me@gmail.com
    name: API Support
  description: API de suivi de comptabilité
  termsOfService: https://example.com/terms/
  title: Accounting API
  version: "1.0"
paths:
...
  /users/:
    post:
      consumes:
      - application/json
      description: Create a new user. A rate limiter prevents from brute force attacks
        (speed 1s, burst 5)
      parameters:
      - description: User payload
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/v1.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/v1.UserResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "429":
          description: Too many requests error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      summary: Create a user
      tags:
      - users
...
  /users/me:
    get:
      consumes:
      - application/json
      description: Retrieve the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.UserResponse'
        "400":
          description: Bad request error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "401":
          description: Unauthorized error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "404":
          description: User not found error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/v1.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Retrieve the authenticated user
      tags:
      - users
schemes:
- http
swagger: "2.0"